Finish implementation of pubsub
The original pubsub implementation didn't actually work
fully, it's supposed to be a callable.

So I've implemented much more of it, enough to mock reasonably.
You do basic things like publish messages, but most everything else
is just a no-op.  Most of what pubsub does doesn't really work well
when mocking anyway, so this is good enough to deal with the basics.

Feel free to suggest more functionality - I only need the basics,
but I'd be happy to do a bit more on this if needed.

Fixes #115